Refactor prescreen-1b to match Harlan design
Harlan's design calls for three different questions on this screen. First, whether the user has a copy of the earnings record. Then the user is either asked the format of the earnings record or asked if they have a social security account. Depending on the answers, manual entry, file upload, file upload with some hints, or help on obtaining earnings records show up.

This commit refactors this page to support these additional questions and displays.

This commit also fixes a previous commit that broke manual entry of earnings.

TODO: Add instructions for signing up on MySocialSecurity. The copy for this was left out of Harlan's mockup